+++ src/main/java/roomescape/global/auth/annotation/RoleRequired.java
@@ -1,4 +1,4 @@
-package roomescape.global.auth;
+package roomescape.global.auth.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
