+++ src/main/java/roomescape/member/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package roomescape.member.unit.repository;
+package roomescape.member.repository;
 
 import java.util.List;
 import java.util.Optional;
