+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.global.auth.AuthenticationPrincipal;
-import roomescape.global.auth.LoginMember;
-import roomescape.global.auth.RoleRequired;
+import roomescape.global.auth.annotation.AuthenticationPrincipal;
+import roomescape.global.auth.annotation.RoleRequired;
+import roomescape.global.auth.dto.LoginMember;
 import roomescape.member.entity.RoleType;
 import roomescape.reservation.dto.request.ReservationRequest.ReservationCreateRequest;
 import roomescape.reservation.dto.request.ReservationRequest.ReservationReadFilteredRequest;
