+++ src/main/java/roomescape/reservation/controller/ReservationTimeController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.global.auth.RoleRequired;
+import roomescape.global.auth.annotation.RoleRequired;
 import roomescape.member.entity.RoleType;
 import roomescape.reservation.dto.request.ReservationTimeRequest.ReservationTimeCreateRequest;
 import roomescape.reservation.dto.response.ReservationTimeResponse.AvailableReservationTimeResponse;
