+++ src/test/java/roomescape/reservation/integration/ReservationTimeIntegrationTest.java
@@ -16,7 +16,7 @@ import roomescape.global.error.exception.ConflictException;
 import roomescape.global.error.exception.NotFoundException;
 import roomescape.member.entity.Member;
 import roomescape.member.entity.RoleType;
-import roomescape.member.unit.repository.MemberRepository;
+import roomescape.member.repository.MemberRepository;
 import roomescape.reservation.dto.request.ReservationTimeRequest.ReservationTimeCreateRequest;
 import roomescape.reservation.entity.Reservation;
 import roomescape.reservation.entity.ReservationTime;
