+++ src/test/java/roomescape/reservation/unit/service/ReservationServiceTest.java
@@ -13,8 +13,8 @@ import roomescape.global.error.exception.ConflictException;
 import roomescape.global.error.exception.NotFoundException;
 import roomescape.member.entity.Member;
 import roomescape.member.entity.RoleType;
+import roomescape.member.repository.MemberRepository;
 import roomescape.member.unit.repository.FakeMemberRepository;
-import roomescape.member.unit.repository.MemberRepository;
 import roomescape.reservation.dto.request.ReservationRequest.ReservationAdminCreateRequest;
 import roomescape.reservation.dto.request.ReservationRequest.ReservationCreateRequest;
 import roomescape.reservation.dto.request.ReservationRequest.ReservationReadFilteredRequest;
