+++ src/test/java/roomescape/theme/acceptance/ThemeAcceptanceTest.java
@@ -14,7 +14,7 @@ import org.springframework.test.annotation.DirtiesContext;
 import roomescape.helper.TestHelper;
 import roomescape.member.entity.Member;
 import roomescape.member.entity.RoleType;
-import roomescape.member.unit.repository.MemberRepository;
+import roomescape.member.repository.MemberRepository;
 import roomescape.theme.dto.request.ThemeRequest.ThemeCreateRequest;
 
 @SpringBootTest(webEnvironment = WebEnvironment.DEFINED_PORT)
